move posthog events to dedicated lib, use enum instead of string for consistency
