+++ src/components/navigation-header/components/sandbox-dropdown/index.tsx
@@ -11,7 +11,7 @@ import { IconChevronRight16 } from '@hashicorp/flight-icons/svg-react/chevron-ri
 import { useRouter } from 'next/router'
 import { useCurrentProduct } from 'contexts'
 import { useInstruqtEmbed } from 'contexts/instruqt-lab'
-import { trackSandboxEvent } from 'lib/analytics'
+import { trackSandboxEvent, SANDBOX_EVENT } from 'lib/posthog-events'
 import useOnClickOutside from 'hooks/use-on-click-outside'
 import useOnEscapeKeyDown from 'hooks/use-on-escape-key-down'
 import useOnFocusOutside from 'hooks/use-on-focus-outside'
@@ -138,7 +138,7 @@ const SandboxDropdown = ({ ariaLabel, label }: SandboxDropdownProps) => {
 	 */
 	const handleLabClick = (lab: SandboxLab) => {
 		openLab(lab.labId)
-		trackSandboxEvent('sandbox_started', {
+		trackSandboxEvent(SANDBOX_EVENT.SANDBOX_STARTED, {
 			labId: lab.labId,
 			page: router.asPath,
 		})
